4d jatkuu sittenkin :D

+++ osa4/blogilista/controllers/blogs.js
@@ -18,24 +18,14 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
-
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+    console.log(request.token)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
       return response.status(401).json({error: 'token invalid'})
     }
     const user = await User.findById(decodedToken.id) 
-
-    console.log('haloo tuleeko tämä mihinkään')
 
     const blog = new Blog({
       title: body.title,
@@ -45,8 +35,6 @@ blogsRouter.post('/', async (request, response) => {
       user: user._id
     })
 
-    console.log('noniin tuleeko tätä', blog.likes)
-
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -54,7 +42,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+
+ const blog = await Blog.findById(request.params.id)
+ if (!blog) {
+  return response.status(401).json({error: 'such blog cannot be found to be removed'})
+ }
+
+ const decodedToken = jwt.verify(request.token, process.env.SECRET)
+ if (!decodedToken.id) {
+   return response.status(401).json({error: 'token invalid'})
+ }
+
+ if (!(blog.user.toString() === decodedToken.id)) {
+  return response.status(400).json({error: 'tokens do not match'})
+ } 
+ await Blog.findByIdAndRemove(request.params.id)
+
   response.status(204).end()
 })
